LPS-84620 journal-service: migrate DDMStructureIndexer related code from JournalArticleIndexer to JournalArticleDDMStructureIndexer
