+++ modules/apps/journal/journal-service/src/main/java/com/liferay/journal/internal/search/JournalArticleIndexer.java
@@ -32,13 +32,7 @@ import com.liferay.journal.util.impl.JournalUtil;
 import com.liferay.petra.string.CharPool;
 import com.liferay.petra.string.StringPool;
 import com.liferay.portal.kernel.dao.orm.ActionableDynamicQuery;
-import com.liferay.portal.kernel.dao.orm.DynamicQuery;
-import com.liferay.portal.kernel.dao.orm.DynamicQueryFactoryUtil;
 import com.liferay.portal.kernel.dao.orm.IndexableActionableDynamicQuery;
-import com.liferay.portal.kernel.dao.orm.ProjectionFactoryUtil;
-import com.liferay.portal.kernel.dao.orm.Property;
-import com.liferay.portal.kernel.dao.orm.PropertyFactoryUtil;
-import com.liferay.portal.kernel.dao.orm.RestrictionsFactoryUtil;
 import com.liferay.portal.kernel.exception.PortalException;
 import com.liferay.portal.kernel.log.Log;
 import com.liferay.portal.kernel.log.LogFactoryUtil;
@@ -47,7 +41,6 @@ import com.liferay.portal.kernel.portlet.PortletRequestModel;
 import com.liferay.portal.kernel.search.BaseIndexer;
 import com.liferay.portal.kernel.search.BooleanClauseOccur;
 import com.liferay.portal.kernel.search.BooleanQuery;
-import com.liferay.portal.kernel.search.DDMStructureIndexer;
 import com.liferay.portal.kernel.search.Document;
 import com.liferay.portal.kernel.search.Field;
 import com.liferay.portal.kernel.search.IndexWriterHelper;
@@ -55,7 +48,6 @@ import com.liferay.portal.kernel.search.Indexer;
 import com.liferay.portal.kernel.search.IndexerRegistry;
 import com.liferay.portal.kernel.search.Query;
 import com.liferay.portal.kernel.search.SearchContext;
-import com.liferay.portal.kernel.search.SearchException;
 import com.liferay.portal.kernel.search.Summary;
 import com.liferay.portal.kernel.search.filter.BooleanFilter;
 import com.liferay.portal.kernel.search.filter.QueryFilter;
@@ -117,8 +109,7 @@ import org.osgi.service.component.annotations.Reference;
  * @author Tibor Lipusz
  */
 @Component(immediate = true, service = Indexer.class)
-public class JournalArticleIndexer
-	extends BaseIndexer<JournalArticle> implements DDMStructureIndexer {
+public class JournalArticleIndexer extends BaseIndexer<JournalArticle> {
 
 	public static final String CLASS_NAME = JournalArticle.class.getName();
 
@@ -288,112 +279,6 @@ public class JournalArticleIndexer
 			if (Validator.isNotNull(expandoAttributes)) {
 				addSearchExpando(searchQuery, searchContext, expandoAttributes);
 			}
-		}
-	}
-
-	@Override
-	public void reindexDDMStructures(List<Long> ddmStructureIds)
-		throws SearchException {
-
-		if (_indexStatusManager.isIndexReadOnly() || !isIndexerEnabled()) {
-			return;
-		}
-
-		try {
-			final String[] ddmStructureKeys =
-				new String[ddmStructureIds.size()];
-
-			for (int i = 0; i < ddmStructureIds.size(); i++) {
-				long ddmStructureId = ddmStructureIds.get(i);
-
-				DDMStructure ddmStructure =
-					_ddmStructureLocalService.getDDMStructure(ddmStructureId);
-
-				ddmStructureKeys[i] = ddmStructure.getStructureKey();
-			}
-
-			final Indexer<JournalArticle> indexer =
-				_indexerRegistry.nullSafeGetIndexer(JournalArticle.class);
-
-			final ActionableDynamicQuery actionableDynamicQuery =
-				_journalArticleResourceLocalService.getActionableDynamicQuery();
-
-			actionableDynamicQuery.setAddCriteriaMethod(
-				new ActionableDynamicQuery.AddCriteriaMethod() {
-
-					@Override
-					public void addCriteria(DynamicQuery dynamicQuery) {
-						Class<?> clazz = getClass();
-
-						DynamicQuery journalArticleDynamicQuery =
-							DynamicQueryFactoryUtil.forClass(
-								JournalArticle.class, "journalArticle",
-								clazz.getClassLoader());
-
-						journalArticleDynamicQuery.setProjection(
-							ProjectionFactoryUtil.property("resourcePrimKey"));
-
-						journalArticleDynamicQuery.add(
-							RestrictionsFactoryUtil.eqProperty(
-								"journalArticle.resourcePrimKey",
-								"this.resourcePrimKey"));
-
-						journalArticleDynamicQuery.add(
-							RestrictionsFactoryUtil.eqProperty(
-								"journalArticle.groupId", "this.groupId"));
-
-						Property ddmStructureKey = PropertyFactoryUtil.forName(
-							"DDMStructureKey");
-
-						journalArticleDynamicQuery.add(
-							ddmStructureKey.in(ddmStructureKeys));
-
-						if (!isIndexAllArticleVersions()) {
-							Property statusProperty =
-								PropertyFactoryUtil.forName("status");
-
-							Integer[] statuses = {
-								WorkflowConstants.STATUS_APPROVED,
-								WorkflowConstants.STATUS_IN_TRASH
-							};
-
-							journalArticleDynamicQuery.add(
-								statusProperty.in(statuses));
-						}
-
-						Property resourcePrimKeyProperty =
-							PropertyFactoryUtil.forName("resourcePrimKey");
-
-						dynamicQuery.add(
-							resourcePrimKeyProperty.in(
-								journalArticleDynamicQuery));
-					}
-
-				});
-			actionableDynamicQuery.setPerformActionMethod(
-				new ActionableDynamicQuery.
-					PerformActionMethod<JournalArticleResource>() {
-
-					@Override
-					public void performAction(JournalArticleResource article)
-						throws PortalException {
-
-						try {
-							indexer.reindex(
-								indexer.getClassName(),
-								article.getResourcePrimKey());
-						}
-						catch (Exception e) {
-							throw new PortalException(e);
-						}
-					}
-
-				});
-
-			actionableDynamicQuery.performActions();
-		}
-		catch (Exception e) {
-			throw new SearchException(e);
 		}
 	}
